chore(core, legacy, storage): refactor flash drivers
[no changelog]
