+++ core/SConscript.boardloader
@@ -50,6 +50,7 @@ SOURCE_MOD += [
     'vendor/trezor-crypto/ed25519-donna/modm-donna-32bit.c',
     'vendor/trezor-crypto/memzero.c',
     'vendor/trezor-crypto/sha2.c',
+    'vendor/trezor-storage/flash_area.c',
 ]
 
 # modtrezorui
