+++ core/SConscript.bootloader_ci
@@ -66,6 +66,7 @@ SOURCE_MOD += [
     'vendor/trezor-crypto/memzero.c',
     'vendor/trezor-crypto/rand.c',
     'vendor/trezor-crypto/sha2.c',
+    'vendor/trezor-storage/flash_area.c',
 ]
 
 # modtrezorui
