+++ core/SConscript.firmware
@@ -58,6 +58,7 @@ SOURCE_MOD += [
     'vendor/trezor-storage/norcow.c',
     'vendor/trezor-storage/storage.c',
     'vendor/trezor-storage/storage_utils.c',
+    'vendor/trezor-storage/flash_area.c',
 ]
 
 # modtrezorcrypto
