+++ core/SConscript.prodtest
@@ -71,6 +71,7 @@ SOURCE_MOD += [
     'vendor/trezor-crypto/secp256k1.c',
     'vendor/trezor-crypto/sha2.c',
     'vendor/trezor-crypto/tls_prf.c',
+    'vendor/trezor-storage/flash_area.c',
 ]
 
 # modtrezorui
