+++ core/SConscript.reflash
@@ -65,6 +65,7 @@ SOURCE_MOD += [
     'vendor/micropython/lib/uzlib/adler32.c',
     'vendor/micropython/lib/uzlib/crc32.c',
     'vendor/micropython/lib/uzlib/tinflate.c',
+    'vendor/trezor-storage/flash_area.c',
 ]
 
 # fonts
