+++ core/SConscript.unix
@@ -59,12 +59,8 @@ SOURCE_MOD += [
     'vendor/trezor-storage/norcow.c',
     'vendor/trezor-storage/storage.c',
     'vendor/trezor-storage/storage_utils.c',
+    'vendor/trezor-storage/flash_area.c',
 ]
-
-if TREZOR_MODEL in ('1', 'T', 'R', 'DISC1'):
-    SOURCE_MOD += [
-        'vendor/trezor-storage/flash_common_f4.c',
-    ]
 
 # modtrezorcrypto
 CCFLAGS_MOD += '-Wno-sequence-point '
@@ -383,6 +379,7 @@ SOURCE_UNIX = [
     'embed/trezorhal/unix/common.c',
     'embed/trezorhal/unix/display-unix.c',
     'embed/trezorhal/unix/flash.c',
+    'embed/trezorhal/unix/flash_otp.c',
     'embed/trezorhal/unix/random_delays.c',
     'embed/trezorhal/unix/rng.c',
     'embed/trezorhal/unix/usb.c',
