+++ core/embed/trezorhal/stm32u5/flash.c
@@ -19,6 +19,7 @@
 
 #include STM32_HAL_H
 
+#include <stdbool.h>
 #include <string.h>
 
 #include "common.h"
@@ -33,22 +34,37 @@
 #define FLASH_STATUS_ALL_FLAGS \
   (FLASH_NSSR_PGSERR | FLASH_NSSR_PGAERR | FLASH_NSSR_WRPERR | FLASH_NSSR_EOP)
 
-uint32_t flash_wait_and_clear_status_flags(void) {
-  while (FLASH->NSSR & FLASH_NSSR_BSY)
-    ;  // wait for all previous flash operations to complete
+static bool flash_sector_is_secure(uint32_t sector) {
+  // We always return true since the entire flash memory is currently secure -
+  // partially through option bytes and partially through  FLASH controller
+  // settings
+  return true;
+}
 
-  uint32_t result =
-      FLASH->NSSR & FLASH_STATUS_ALL_FLAGS;  // get the current status flags
-  FLASH->NSSR |= FLASH_STATUS_ALL_FLAGS;     // clear all status flags
+const void *flash_get_address(uint16_t sector, uint32_t offset, uint32_t size) {
+  if (sector >= FLASH_SECTOR_COUNT) {
+    return NULL;
+  }
 
-#if defined(__ARM_FEATURE_CMSE) && (__ARM_FEATURE_CMSE == 3U)
-  while (FLASH->SECSR & FLASH_SECSR_BSY)
-    ;  // wait for all previous flash operations to complete
-  result |=
-      FLASH->SECSR & FLASH_STATUS_ALL_FLAGS;  // get the current status flags
-  FLASH->SECSR |= FLASH_STATUS_ALL_FLAGS;     // clear all status flags
-#endif
-  return result;
+  if (offset + size > FLASH_PAGE_SIZE) {
+    return NULL;
+  }
+
+  uint32_t base_addr = flash_sector_is_secure(sector) ? FLASH_SEC_START_ADDRESS
+                                                      : FLASH_START_ADDRESS;
+
+  return (const void *)(base_addr + FLASH_PAGE_SIZE * sector + offset);
+}
+
+uint32_t flash_sector_size(uint16_t first_sector, uint16_t sector_count) {
+  if (first_sector + sector_count > FLASH_SECTOR_COUNT) {
+    return 0;
+  }
+  return FLASH_PAGE_SIZE * sector_count;
+}
+
+uint16_t flash_sector_find(uint16_t first_sector, uint32_t offset) {
+  return first_sector + offset / FLASH_PAGE_SIZE;
 }
 
 secbool flash_unlock_write(void) {
@@ -65,297 +81,53 @@ secbool flash_lock_write(void) {
   return sectrue;
 }
 
-const void *flash_get_address(uint16_t sector, uint32_t offset, uint32_t size,
-                              bool secure_area) {
+secbool flash_sector_erase(uint16_t sector) {
   if (sector >= FLASH_SECTOR_COUNT) {
-    return NULL;
-  }
-  uint32_t base_addr = FLASH_START_ADDRESS;
-
-  if (secure_area) {
-    base_addr = FLASH_SEC_START_ADDRESS;
+    return secfalse;
   }
 
-  const uint32_t addr = base_addr + (FLASH_PAGE_SIZE * sector) + offset;
-  const uint32_t next = base_addr + (FLASH_PAGE_SIZE * (sector + 1));
-  if (addr + size > next) {
-    return NULL;
-  }
-  return (const void *)addr;
-}
+  FLASH_EraseInitTypeDef EraseInitStruct = {
+      .TypeErase = FLASH_TYPEERASE_PAGES_NS,
+      .Banks = FLASH_BANK_1,
+      .Page = sector,
+      .NbPages = 1,
+  };
 
-uint32_t flash_sector_size(uint16_t sector) {
-  if (sector >= FLASH_SECTOR_COUNT) {
-    return 0;
-  }
-  return FLASH_PAGE_SIZE;
-}
-
-uint32_t flash_subarea_get_size(const flash_subarea_t *sub) {
-  return FLASH_PAGE_SIZE * sub->num_sectors;
-}
-
-uint32_t flash_area_get_size(const flash_area_t *area) {
-  uint32_t size = 0;
-  for (int i = 0; i < area->num_subareas; i++) {
-    size += flash_subarea_get_size(&area->subarea[i]);
-  }
-  return size;
-}
-
-uint16_t flash_total_sectors(const flash_area_t *area) {
-  uint16_t total = 0;
-  for (int i = 0; i < area->num_subareas; i++) {
-    total += area->subarea[i].num_sectors;
-  }
-  return total;
-}
-
-const void *flash_area_get_address(const flash_area_t *area, uint32_t offset,
-                                   uint32_t size) {
-  uint32_t tmp_offset = offset;
-
-  for (int i = 0; i < area->num_subareas; i++) {
-    uint16_t sector = area->subarea[i].first_sector;
-
-    if (tmp_offset >= flash_subarea_get_size(area->subarea)) {
-      tmp_offset -= flash_subarea_get_size(area->subarea);
-      continue;
-    }
-
-    uint32_t base_addr = FLASH_START_ADDRESS;
-
-    if (area->secure_area) {
-      base_addr = FLASH_SEC_START_ADDRESS;
-    }
-
-    const uint32_t addr = base_addr + (FLASH_PAGE_SIZE * sector) + tmp_offset;
-    const uint32_t area_end =
-        base_addr + (FLASH_PAGE_SIZE * (area->subarea[i].first_sector +
-                                        area->subarea[i].num_sectors));
-    if (addr + size > area_end) {
-      return NULL;
-    }
-    return (const void *)addr;
-  }
-  return NULL;
-}
-
-int32_t flash_get_sector_num(const flash_area_t *area,
-                             uint32_t sector_inner_num) {
-  uint16_t sector = 0;
-  uint16_t remaining = sector_inner_num;
-  for (int i = 0; i < area->num_subareas; i++) {
-    if (remaining < area->subarea[i].num_sectors) {
-      sector = area->subarea[i].first_sector + remaining;
-      return sector;
-    } else {
-      remaining -= area->subarea[i].num_sectors;
-    }
+  if (sector >= 256) {
+    EraseInitStruct.Banks = FLASH_BANK_2;
+    EraseInitStruct.Page = sector - 256;
   }
 
-  return -1;
-}
-
-secbool flash_area_erase(const flash_area_t *area,
-                         void (*progress)(int pos, int len)) {
-  ensure(flash_unlock_write(), NULL);
-  FLASH_EraseInitTypeDef EraseInitStruct;
-  EraseInitStruct.TypeErase = FLASH_TYPEERASE_PAGES;
-  EraseInitStruct.NbPages = 1;
-
-  int total_pages = 0;
-  int done_pages = 0;
-  for (int i = 0; i < area->num_subareas; i++) {
-    total_pages += area->subarea[i].num_sectors;
+  if (flash_sector_is_secure(sector)) {
+    EraseInitStruct.TypeErase = FLASH_TYPEERASE_PAGES;
   }
 
-  if (progress) {
-    progress(0, total_pages);
+  uint32_t sector_error = 0;
+
+  if (HAL_FLASHEx_Erase(&EraseInitStruct, &sector_error) != HAL_OK) {
+    return secfalse;
   }
 
-  for (int s = 0; s < area->num_subareas; s++) {
-    for (int i = 0; i < area->subarea[s].num_sectors; i++) {
-      int page = area->subarea[s].first_sector + i;
-      uint32_t sec_start = 0;
-      uint32_t sec_end = 0;
-      uint32_t page_idx = 0;
-      bool secure = area->secure_area;
-      if (page >= 256) {
-        EraseInitStruct.Banks = FLASH_BANK_2;
-        page_idx = page - 256;
-        sec_start = (FLASH->SECWM2R1 & FLASH_SECWM1R1_SECWM1_PSTRT) >>
-                    FLASH_SECWM1R1_SECWM1_PSTRT_Pos;
-        sec_end = (FLASH->SECWM2R1 & FLASH_SECWM1R1_SECWM1_PEND) >>
-                  FLASH_SECWM1R1_SECWM1_PEND_Pos;
-      } else {
-        EraseInitStruct.Banks = FLASH_BANK_1;
-        page_idx = page;
-        sec_start = (FLASH->SECWM1R1 & FLASH_SECWM1R1_SECWM1_PSTRT) >>
-                    FLASH_SECWM1R1_SECWM1_PSTRT_Pos;
-        sec_end = (FLASH->SECWM1R1 & FLASH_SECWM1R1_SECWM1_PEND) >>
-                  FLASH_SECWM1R1_SECWM1_PEND_Pos;
-      }
+  // check whether the sector was really deleted (contains only 0xFF)
+  const uint32_t *sector_start =
+      (const uint32_t *)flash_get_address(sector, 0, 0);
 
-      if (page_idx < sec_start || page_idx > sec_end) {
-        EraseInitStruct.TypeErase = FLASH_TYPEERASE_PAGES_NS;
-        secure = false;
-      } else {
-        EraseInitStruct.TypeErase = FLASH_TYPEERASE_PAGES;
-      }
+  const uint32_t *sector_end =
+      sector_start + flash_sector_size(sector, 1) / sizeof(uint32_t);
 
-      EraseInitStruct.Page = page_idx;
-      uint32_t SectorError;
-      if (HAL_FLASHEx_Erase(&EraseInitStruct, &SectorError) != HAL_OK) {
-        ensure(flash_lock_write(), NULL);
-        return secfalse;
-      }
-      // check whether the sector was really deleted (contains only 0xFF)
-      const uint32_t addr_start =
-          (uint32_t)flash_get_address(page, 0, 0, secure);
-      const uint32_t addr_end =
-          (uint32_t)flash_get_address(page + 1, 0, 0, secure);
-
-      for (uint32_t addr = addr_start; addr < addr_end; addr += 4) {
-        if (*((const uint32_t *)addr) != 0xFFFFFFFF) {
-          ensure(flash_lock_write(), NULL);
-          return secfalse;
-        }
-      }
-      done_pages++;
-      if (progress && done_pages % 16 == 0) {
-        progress(done_pages, total_pages);
-      }
-    }
-  }
-  ensure(flash_lock_write(), NULL);
-  return sectrue;
-}
-
-secbool flash_area_erase_bulk(const flash_area_t *areas, int count,
-                              void (*progress)(int pos, int len)) {
-  ensure(flash_unlock_write(), NULL);
-  FLASH_EraseInitTypeDef EraseInitStruct;
-  EraseInitStruct.TypeErase = FLASH_TYPEERASE_PAGES;
-  EraseInitStruct.NbPages = 1;
-
-  int total_pages = 0;
-  int done_pages = 0;
-  for (int c = 0; c < count; c++) {
-    for (int i = 0; i < areas[c].num_subareas; i++) {
-      total_pages += areas[c].subarea[i].num_sectors;
-    }
-  }
-
-  if (progress) {
-    progress(0, total_pages);
-  }
-
-  for (int c = 0; c < count; c++) {
-    for (int s = 0; s < areas[c].num_subareas; s++) {
-      for (int i = 0; i < areas[c].subarea[s].num_sectors; i++) {
-        int page = areas[c].subarea[s].first_sector + i;
-        if (page >= 256) {
-          EraseInitStruct.Banks = FLASH_BANK_2;
-        } else {
-          EraseInitStruct.Banks = FLASH_BANK_1;
-        }
-        EraseInitStruct.Page = page;
-        uint32_t SectorError;
-        if (HAL_FLASHEx_Erase(&EraseInitStruct, &SectorError) != HAL_OK) {
-          ensure(flash_lock_write(), NULL);
-          return secfalse;
-        }
-        // check whether the sector was really deleted (contains only 0xFF)
-        const uint32_t addr_start =
-            (uint32_t)flash_get_address(page, 0, 0, areas[c].secure_area);
-        const uint32_t addr_end =
-            (uint32_t)flash_get_address(page + 1, 0, 0, areas[c].secure_area);
-
-        for (uint32_t addr = addr_start; addr < addr_end; addr += 4) {
-          if (*((const uint32_t *)addr) != 0xFFFFFFFF) {
-            ensure(flash_lock_write(), NULL);
-            return secfalse;
-          }
-        }
-        done_pages++;
-        if (progress) {
-          progress(done_pages, total_pages);
-        }
-      }
-    }
-  }
-  ensure(flash_lock_write(), NULL);
-  return sectrue;
-}
-
-secbool flash_area_erase_partial(const flash_area_t *area, uint32_t offset,
-                                 uint32_t *bytes_erased) {
-  ensure(flash_unlock_write(), NULL);
-  FLASH_EraseInitTypeDef EraseInitStruct;
-  EraseInitStruct.TypeErase = FLASH_TYPEERASE_PAGES;
-  EraseInitStruct.NbPages = 1;
-
-  *bytes_erased = 0;
-  int32_t page = flash_get_sector_num(area, offset / FLASH_PAGE_SIZE);
-
-  if (page >= 0 && page < FLASH_SECTOR_COUNT) {
-    uint32_t sec_start = 0;
-    uint32_t sec_end = 0;
-    uint32_t page_idx = 0;
-    bool secure = area->secure_area;
-    if (page >= 256) {
-      EraseInitStruct.Banks = FLASH_BANK_2;
-      page_idx = page - 256;
-      sec_start = (FLASH->SECWM2R1 & FLASH_SECWM1R1_SECWM1_PSTRT) >>
-                  FLASH_SECWM1R1_SECWM1_PSTRT_Pos;
-      sec_end = (FLASH->SECWM2R1 & FLASH_SECWM1R1_SECWM1_PEND) >>
-                FLASH_SECWM1R1_SECWM1_PEND_Pos;
-    } else {
-      EraseInitStruct.Banks = FLASH_BANK_1;
-      page_idx = page;
-      sec_start = (FLASH->SECWM1R1 & FLASH_SECWM1R1_SECWM1_PSTRT) >>
-                  FLASH_SECWM1R1_SECWM1_PSTRT_Pos;
-      sec_end = (FLASH->SECWM1R1 & FLASH_SECWM1R1_SECWM1_PEND) >>
-                FLASH_SECWM1R1_SECWM1_PEND_Pos;
-    }
-
-    if (page_idx < sec_start || page_idx > sec_end) {
-      EraseInitStruct.TypeErase = FLASH_TYPEERASE_PAGES_NS;
-      secure = false;
-    } else {
-      EraseInitStruct.TypeErase = FLASH_TYPEERASE_PAGES;
-    }
-
-    EraseInitStruct.Page = page_idx;
-    uint32_t SectorError;
-    if (HAL_FLASHEx_Erase(&EraseInitStruct, &SectorError) != HAL_OK) {
-      ensure(flash_lock_write(), NULL);
+  for (const uint32_t *addr = sector_start; addr < sector_end; addr++) {
+    if (*addr != 0xFFFFFFFF) {
       return secfalse;
     }
-    // check whether the sector was really deleted (contains only 0xFF)
-    const uint32_t addr_start = (uint32_t)flash_get_address(page, 0, 0, secure);
-    const uint32_t addr_end =
-        (uint32_t)flash_get_address(page + 1, 0, 0, secure);
-
-    for (uint32_t addr = addr_start; addr < addr_end; addr += 4) {
-      if (*((const uint32_t *)addr) != 0xFFFFFFFF) {
-        ensure(flash_lock_write(), NULL);
-        return secfalse;
-      }
-    }
-    ensure(flash_lock_write(), NULL);
-    *bytes_erased = FLASH_PAGE_SIZE;
-    return sectrue;
   }
-  ensure(flash_lock_write(), NULL);
+
   return sectrue;
 }
 
 secbool flash_write_quadword(uint16_t sector, uint32_t offset,
-                             const uint32_t *data, bool secure_area) {
-  uint32_t address = (uint32_t)flash_get_address(
-      sector, offset, 4 * sizeof(uint32_t), secure_area);
+                             const uint32_t *data) {
+  uint32_t address =
+      (uint32_t)flash_get_address(sector, offset, 4 * sizeof(uint32_t));
   if (address == 0) {
     return secfalse;
   }
@@ -394,9 +166,9 @@ secbool flash_write_quadword(uint16_t sector, uint32_t offset,
 }
 
 secbool flash_write_burst(uint16_t sector, uint32_t offset,
-                          const uint32_t *data, bool secure_area) {
-  uint32_t address = (uint32_t)flash_get_address(
-      sector, offset, 8 * 4 * sizeof(uint32_t), secure_area);
+                          const uint32_t *data) {
+  uint32_t address =
+      (uint32_t)flash_get_address(sector, offset, 8 * 4 * sizeof(uint32_t));
   if (address == 0) {
     return secfalse;
   }
@@ -433,132 +205,6 @@ secbool flash_write_burst(uint16_t sector, uint32_t offset,
     }
   }
   return sectrue;
-}
-
-secbool flash_area_write_quadword(const flash_area_t *area, uint32_t offset,
-                                  const uint32_t *data) {
-  uint32_t tmp_offset = offset;
-  for (int i = 0; i < area->num_subareas; i++) {
-    uint16_t sector = area->subarea[i].first_sector;
-
-    uint32_t sub_size = flash_subarea_get_size(&area->subarea[i]);
-    if (tmp_offset >= sub_size) {
-      tmp_offset -= sub_size;
-      continue;
-    }
-
-    // in correct subarea
-    for (int s = 0; s < area->subarea[i].num_sectors; s++) {
-      const uint32_t sector_size = flash_sector_size(sector);
-      if (tmp_offset >= sector_size) {
-        tmp_offset -= sector_size;
-        sector++;
-
-        if (s == area->subarea[i].num_sectors - 1) {
-          return secfalse;
-        }
-        continue;
-      }
-      // in correct sector
-      return flash_write_quadword(sector, tmp_offset, data, area->secure_area);
-    }
-  }
-  return secfalse;
-}
-
-secbool flash_area_write_burst(const flash_area_t *area, uint32_t offset,
-                               const uint32_t *data) {
-  uint32_t tmp_offset = offset;
-  for (int i = 0; i < area->num_subareas; i++) {
-    uint16_t sector = area->subarea[i].first_sector;
-
-    uint32_t sub_size = flash_subarea_get_size(&area->subarea[i]);
-    if (tmp_offset >= sub_size) {
-      tmp_offset -= sub_size;
-      continue;
-    }
-
-    // in correct subarea
-    for (int s = 0; s < area->subarea[i].num_sectors; s++) {
-      const uint32_t sector_size = flash_sector_size(sector);
-      if (tmp_offset >= sector_size) {
-        tmp_offset -= sector_size;
-        sector++;
-
-        if (s == area->subarea[i].num_sectors - 1) {
-          return secfalse;
-        }
-        continue;
-      }
-      // in correct sector
-      return flash_write_burst(sector, tmp_offset, data, area->secure_area);
-    }
-  }
-  return secfalse;
-}
-
-secbool flash_otp_read(uint8_t block, uint8_t offset, uint8_t *data,
-                       uint8_t datalen) {
-  if (block >= FLASH_OTP_NUM_BLOCKS ||
-      offset + datalen > FLASH_OTP_BLOCK_SIZE) {
-    return secfalse;
-  }
-  for (uint8_t i = 0; i < datalen; i++) {
-    data[i] = *(__IO uint8_t *)(FLASH_OTP_BASE + block * FLASH_OTP_BLOCK_SIZE +
-                                offset + i);
-  }
-  return sectrue;
-}
-
-secbool flash_otp_write(uint8_t block, uint8_t offset, const uint8_t *data,
-                        uint8_t datalen) {
-  if (datalen % 16 != 0) {
-    return secfalse;
-  }
-  if (block >= FLASH_OTP_NUM_BLOCKS ||
-      offset + datalen > FLASH_OTP_BLOCK_SIZE) {
-    return secfalse;
-  }
-  ensure(flash_unlock_write(), NULL);
-  for (uint8_t i = 0; i < datalen; i++) {
-    uint32_t address =
-        FLASH_OTP_BASE + block * FLASH_OTP_BLOCK_SIZE + offset + i;
-    ensure(sectrue * (HAL_OK == HAL_FLASH_Program(FLASH_TYPEPROGRAM_QUADWORD,
-                                                  address, (uint32_t)data)),
-           NULL);
-  }
-  ensure(flash_lock_write(), NULL);
-  return sectrue;
-}
-
-secbool flash_otp_lock(uint8_t block) {
-  // check that all quadwords in the block have been written to
-  volatile uint8_t *addr =
-      (__IO uint8_t *)(FLASH_OTP_BASE + block * FLASH_OTP_BLOCK_SIZE);
-
-  secbool qw_locked = secfalse;
-  for (uint8_t i = 0; i < FLASH_OTP_BLOCK_SIZE; i++) {
-    if (addr[i] != 0xFF) {
-      qw_locked = sectrue;
-    }
-    if (i % 16 == 15 && qw_locked == secfalse) {
-      return secfalse;
-    }
-  }
-  return sectrue;
-}
-
-secbool flash_otp_is_locked(uint8_t block) {
-  // considering block locked if any quadword in the block is non-0xFF
-  volatile uint8_t *addr =
-      (__IO uint8_t *)(FLASH_OTP_BASE + block * FLASH_OTP_BLOCK_SIZE);
-
-  for (uint8_t i = 0; i < FLASH_OTP_BLOCK_SIZE; i++) {
-    if (addr[i] != 0xFF) {
-      return sectrue;
-    }
-  }
-  return secfalse;
 }
 
 secbool flash_write_block(uint16_t sector, uint32_t offset,
