+++ core/embed/trezorhal/stm32u5/secret.c
@@ -1,4 +1,5 @@
 #include "secret.h"
+#include <stdbool.h>
 #include <string.h>
 #include "common.h"
 #include "flash.h"
