+++ core/site_scons/boards/stm32f4_common.py
@@ -44,6 +44,7 @@ def stm32f4_common_files(env, defines, sources, paths):
         "embed/trezorhal/stm32f4/common.c",
         "embed/trezorhal/stm32f4/fault_handlers.c",
         "embed/trezorhal/stm32f4/flash.c",
+        "embed/trezorhal/stm32f4/flash_otp.c",
         "embed/trezorhal/stm32f4/lowlevel.c",
         "embed/trezorhal/stm32f4/mpu.c",
         "embed/trezorhal/stm32f4/platform.c",
@@ -53,7 +54,6 @@ def stm32f4_common_files(env, defines, sources, paths):
         "embed/trezorhal/stm32f4/rng.c",
         "embed/trezorhal/stm32f4/vectortable.s",
         "embed/trezorhal/stm32f4/translations.c",
-        "vendor/trezor-storage/flash_common_f4.c",
     ]
 
     # boardloader needs separate assembler for some function unencumbered by various FW+bootloader hacks
