+++ legacy/Makefile
@@ -20,7 +20,7 @@ OBJS += usb_standard.o
 OBJS += util.o
 OBJS += webusb.o
 OBJS += winusb.o
-OBJS += vendor/trezor-storage/flash_common_f4.o
+OBJS += vendor/trezor-storage/flash_area.o
 
 libtrezor.a: $(OBJS)
