+++ storage/tests/c/Makefile
@@ -14,7 +14,7 @@ SRC  = storage/tests/c/flash.c
 SRC += storage/tests/c/common.c
 SRC += storage/tests/c/random_delays.c
 SRC += storage/tests/c/test_layout.c
-SRC += storage/flash_common_f4.c
+SRC += storage/flash_area.c
 SRC += storage/storage.c
 SRC += storage/storage_utils.c
 SRC += storage/norcow.c
